Get runner auto-updates working in OpenShift
Because of a quirk in how the auto-update copies files, this requires
a hack in entrypoint.sh to ensure the runner's linux user inside the
container owns the .sh files of the runner.

It also requires switching from `run.sh` to `bin/runsvc.sh` as the
latter handles restarts of the runner when it auto-updates.